Supported import/export of arguments also for tasks without schema
