+++ src/routes/projects/[projectId]/workflows/experimental/[workflowId]/+page.svelte
@@ -902,6 +902,7 @@
 													workflowTaskId={selectedWorkflowTask.id}
 													argumentsSchema={selectedWorkflowTask.task.args_schema}
 													argumentsSchemaVersion={selectedWorkflowTask.task.args_schema_version}
+													taskName={selectedWorkflowTask.task.name}
 													args={selectedWorkflowTask.args}
 													bind:saveChanges={saveArgumentsChanges}
 													bind:validSchema={argsSchemaValid}
@@ -911,8 +912,7 @@
 											{:else}
 												<ArgumentForm
 													workflowId={workflow.id}
-													workflowTaskId={selectedWorkflowTask.id}
-													workflowTaskArgs={selectedWorkflowTask.args}
+													workflowTask={selectedWorkflowTask}
 												/>
 											{/if}
 										{/key}
